make integer

+++ src/estimator.py
@@ -26,8 +26,8 @@ def estimator(data):
 #   currentlyInfected["impact"]["severeCasesByRequestedTime"] = 0.15 * currentlyInfected["impact"]["infectionsByRequestedTime"]
 #   currentlyInfected["severeImpact"]["severeCasesByRequestedTime"] = 0.15 * currentlyInfected["severeImpact"]["infectionsByRequestedTime"]
 
-  currentlyInfected["impact"]["severeCasesByRequestedTime"] = 0.15 * currentlyInfected["impact"]["infectionsByRequestedTime"] 
-  currentlyInfected["severeImpact"]["severeCasesByRequestedTime"] = 0.15 * currentlyInfected["severeImpact"]["infectionsByRequestedTime"] 
+  currentlyInfected["impact"]["severeCasesByRequestedTime"] = int(0.15 * currentlyInfected["impact"]["infectionsByRequestedTime"])
+  currentlyInfected["severeImpact"]["severeCasesByRequestedTime"] = int(0.15 * currentlyInfected["severeImpact"]["infectionsByRequestedTime"])
   
   currentlyInfected["impact"]["hospitalBedsByRequestedTime"] = int(data['totalHospit)alBeds']*0.35 - currentlyInfected["impact"]["severeCasesByRequestedTime"])
   currentlyInfected["severeImpact"]["hospitalBedsByRequestedTime"] = int(data['totalHospitalBeds']*0.35 - currentlyInfected["severeImpact"]["severeCasesByRequestedTime"])
